requirements.txt
asgiref==3.3.1
binary==1.0.0
click==7.1.2
Django==3.1.3
django-leaflet==0.27.1
django-numpy==1.0.1
ez-setup==0.9
GDAL @ file:///E:/FYP%20code/GDAL-3.1.4-cp39-cp39-win_amd64.whl
numpy==1.19.3
pip-tools==5.4.0
psycopg2==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
